requirements.txt
alembic==1.8.1
autopep8==2.0.0
blinker==1.5
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-MySQLdb==1.0.1
greenlet==1.1.3.post0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycodestyle==2.9.1
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2